+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 Bug fixes:
 
+  - [dl] Fixed off-by-one error with plain text goto definition
   - [fw] Ensure inotify is stopped before shutting down
   - [wr] properly deconstruct array in foreach
   - [lsp] import unresolved classes refactoring: Ensure only unique names are shown when asking user to select an import candidate
@@ -22,6 +23,7 @@ Bug fixes:
 
 Improvements:
 
+  - [wr] handle static properly #967
   - [ls] include list of diagnostic providers in status report
   - [ct] add retutn type to generated method if it would immediately return
   - [wr] support array [] addition operator
@@ -46,6 +48,7 @@ Improvements:
 
 Features:
 
+  - [wr] Handle constant glob to union types (`@return Foo::BAR_*`).
   - [lsp] show class category in offset hover info 
   - [lsp] jump to types in a union type
   - [wr] Type combination